ecourseapi/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cloudinary==1.44.1
Django==6.0
django-ckeditor==6.7.3
django-js-asset==3.1.2
PyMySQL==1.1.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
